credit_system/requirements.txt
lightgbm>=4.0.0
shap>=0.44.0
scikit-learn>=1.3.0
fairlearn>=0.10.0
pandas>=2.0.0
numpy>=1.24.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
joblib>=1.3.0

# ── KYC OCR stack ──────────────────────────────────────────
easyocr>=1.7.1           # OCR engine (EasyOCR)
PyMuPDF>=1.24.0          # PDF → image rasteriser (imported as fitz)
python-multipart>=0.0.9  # FastAPI file-upload support
PyPDF2>=3.0.0
Pillow>=10.0.0
pytesseract>=0.3.10
